Add in a C solution
Develop library functions for loading the entire array of nums into
memory. This uses an ever expanding array to amortize the expansion
time.

The algorithm here is more optimized as well, using an explicitly
branchless style and an optimized version of the three window.

As like the rust solution, the filename is made based on the executable
name, meant to be run directly in the `c/` directory.