Day 13 in Kotlin
Part 1, split out the lines into pairs, find the first instruction, and
then do the fold based on the instruction of the fold, and then print
out thes size of the set created.

Part 2, parse out the initial points and the instructions, and then fold
the instructions as a set of folds to undertake like part 1. Find the
maximum size of the grid and just print out the resultant grid, since
having to parse this by hand doesn't really work well, and we don't need
to parse the text via computer, just print it out as part of the
solution.