Day 14 in Kotlin
Similar to Day 6 but with pairs of characters instead of int states.

Needed a little bit of Reddit help for the counting, to make sure I
don't double count, and that worked correctly when thought about.

Leaving in the brute force part 1 solution, since that is the easier to
come up with.