Day 15 in Kotlin
While both parts should have the same actual solution (the only
difference is the tiling), I opted to solve the first solution using a
DP floodfill of the board. This also works on the example problem for
solution 2, but this does not work for the full problem, and in place of
a DP floodfill, I use a modified Dijsktra's which terminates when the
corner is found to put in the array.