Day 16 in Kotlin
Longer solution tonight.
Originally tried to operate directly on the Hex Values, but that did not
work well. Use polymorphism to overload a value function for part 2.
Divide and conquer the packet parsing, parsing the header and the data,
and passing the remaning data out to the caller to parse further.